Create codeflix-churnrate-project
This is my query to calculate the churn rate between these two groups of users (segment 87 and segment 30), in between the 1st of January to the 31st of March. >> The segment with the lowest churn rate is segment 30, with on average a churn rate 4 times inferior to segment 87 churn rate. This information would be very valuable to the marketing team to understand that segment 87 might not be as of a good investment as segment 30

Four months into launching Codeflix, management asks you to look into subscription churn rates. The marketing department is particularly interested in how the churn compares between two segments of users. They provide you with a dataset containing subscription data for users who were acquired through two distinct channels. Codeflix requires a minimum subscription length of 31 days, so a user can never start and end their subscription in the same month.